refactor(assay-lua): plan 15 slice 4 — assay.engine.{core,auth,workflow} client trio
Replace the legacy assay.workflow + assay.auth + assay.engine modules
with a unified `assay.engine` umbrella backed by three submodule
clients. Mirrors the assay.ory shape exactly so the developer
experience is consistent across the two stdlib namespaces.

  local engine = require("assay.engine").connect("http://localhost:8420", "admin-key")
  engine.core:info()           -- engine identity (public)
  engine.auth:login(email, pw) -- session login
  engine.workflow:start({...}) -- start a workflow
  engine.workflow:listen({...}) -- become a worker

Submodule structure (matches existing per-vendor namespacing):

  assay.engine                — umbrella with `connect()` + re-exports
  assay.engine.core           — engine-core admin (info, modules,
                                instances, audit, config)
  assay.engine.auth           — full auth surface: login/logout/whoami,
                                passkey, federated OIDC, biscuit,
                                Zanzibar, admin (users / sessions /
                                OIDC clients / OIDC upstream / JWKS /
                                audit), and the public OIDC provider
                                spec endpoints (discovery, token,
                                userinfo, revoke, introspect)
  assay.engine.workflow       — workflow CRUD + signals + cancel +
                                state + child workflows + schedules +
                                namespaces + workers + queues, plus
                                `register_workflow` / `register_activity`
                                / `listen` for worker mode
  assay.engine.workflow.ctx   — replay machinery (the workflow ctx
                                factory used during worker dispatch)
  assay.engine.workflow.worker — poll loop + task handlers

Pattern shift: the legacy `assay.workflow` was module-scoped (one
implicit connection per process); the new `engine.workflow.client(opts)`
returns a stateful client. Multiple clients can target different
engines from the same script.

URLs: every call uses the new `/api/v1/engine/{core,auth,workflow}/*`
surface from slices 1 + 2. OIDC spec endpoints stay under `/auth/*`.

DELETE crates/assay/stdlib/auth.lua (511 LOC).
DELETE crates/assay/stdlib/workflow.lua (503 LOC).
DELETE crates/assay/stdlib/workflow/{api,ctx,listen,task}.lua (782 LOC).

Net: -2000 LOC of old client + -42 LOC of dead engine.lua replaced
with the new client trio (~1250 LOC).

Tests stay red against the partial state per the plan-15 strategy
— slice 5 brings up the Lua test suite + init.lua against the new
client.
